feat(dasclaw_runtime): port SecretsCrypto (AES-256-GCM + HKDF) to shared crate (F3.2 phase 2 PR 4b of 6)
Stacked on PR 4a (#646). Closes part of #641.

## 背景

F3.2 phase 2 把 secrets 子系统从 ironclaw 搬到 dasclaw_runtime,分成
4a/4b/4c 三步。本 PR 是第 2 步:加密层。

PR 4a 已经把 `SecretsStore` trait 和 7 个类型(`Secret`/`SecretRef`/
`DecryptedSecret`/`SecretError`/`CreateSecretParams`/`CredentialLocation`/
`CredentialMapping`)搬过去了。本 PR 继续搬 `SecretsCrypto`(AES-256-GCM
+ HKDF-SHA256 per-secret key derivation),它是 pure-crypto,无 OS 依赖、
无品牌耦合,适合直接搬到框架 crate。

## 改动范围

新增 `crates/dasclaw_runtime/src/secrets/crypto.rs`(~370 行):
- `SecretsCrypto::new(SecretString) -> Result<Self, SecretError>` —— 校验主
  密钥长度 ≥ 32 字节
- `encrypt(plaintext) -> (Vec<u8>, Vec<u8>)` —— 每次生成随机 salt,返回
  `nonce || ciphertext || tag` 和 salt
- `decrypt(encrypted, salt) -> DecryptedSecret` —— 长度校验 +
  AES-256-GCM 认证解密
- `derive_key(salt)` —— HKDF-SHA256 派生 32 字节子密钥,info 串保留
  `"near-agent-secrets-v1"`(兼容已加密数据)
- `Debug` 实现把 master_key 渲染为 `[REDACTED]`

依赖(在 `crates/dasclaw_runtime/Cargo.toml`):
- `aes-gcm = "0.10"`
- `hkdf = "0.12"`
- `sha2 = "0.10"`
- `rand = "0.8"`

`crates/dasclaw_runtime/src/secrets/mod.rs`:
- 新增 `pub mod crypto;` + `pub use crypto::SecretsCrypto;`
- 文档更新:说明 keychain 还没搬(留给 4b' 单独处理参数化),具体后端
  和 agent_provider 留给 4c

`desktop-client/ironclaw/src/secrets/crypto.rs` 改成 7 行 shim:
```rust
pub use dasclaw_runtime::secrets::crypto::SecretsCrypto;
```
所有现存调用方(20 处)零修改。`crate::secrets::crypto::SecretsCrypto`
和 `crate::secrets::SecretsCrypto` 两种路径都仍能解析。

## 测试

18 个 `req_dasclaw_runtime_secrets_crypto_*` 测试,涵盖:
- 加解密往返(含 1MB 大块、空 plaintext、Unicode)
- 安全失败路径(篡改密文 / 篡改 nonce / 错 salt / 错主密钥 → 全部
  Fail-Safe 拒绝,不会 Fail-Open)
- 主密钥隔离(同明文用不同主密钥加密后无法交叉解密)
- 主密钥长度边界(31 字节拒绝、32 字节通过、128 字节通过)
- salt 性质(长度正确、非全 0、两次生成互异)
- 密文结构(`len == NONCE_SIZE + plaintext_len + TAG_SIZE`)
- 截断密文(< NONCE+TAG)立即拒绝
- `Debug` 输出不泄漏主密钥

ironclaw 测试 fixture(`testing::credentials::TEST_CRYPTO_KEY`)只在
ironclaw 内部使用;新 crate 内的测试自带一份本地常量,故意不依赖
host fixture(保持框架 crate 的独立性)。

## 非目标(明确说出,避免补丁式扩张)

- **keychain.rs 不在本 PR**:它把 `SERVICE_NAME = "ironclaw"` 硬编
  码,直接搬会把宿主品牌泄漏进框架 crate。需要先做 `service_name`
  参数化设计(把常量改成 `KeychainConfig { service_name: &str }`),
  这是行为改动,留给 PR 4b' 单独审查。
- **Postgres / LibSQL / InMemory 三个具体 `SecretsStore` impl 不在本
  PR**:留给 4c。
- **agent_provider.rs(`SecretsProvider` 适配器)不在本 PR**:留给 4c。
- **`SecretError` 没新增变体**:加密层用的全部错误码(`InvalidMasterKey`
  / `EncryptionFailed` / `DecryptionFailed`)都在 PR 4a 已经搬过的
  `SecretError` 里。
- **加密格式版本号 `"near-agent-secrets-v1"` 保持不变**:任何修改都会让
  已加密数据无法解密,留给后续 ADR 讨论。

## 关联 ADR

- `docs/plans/architecture-refactor/adr-153-headless-agent-framework-draft.md`
  —— 无头 agent 框架草稿,本 PR 是其下 "shared crate 化 secrets" 那一节
  的第 2 步

## 验证(命令 + 结果)

```bash
cargo check -p dasclaw_runtime --tests        # 12.93s OK
cargo check -p dasclaw --tests                 # 5m40s OK(ironclaw 全量编译)
cargo nextest run -p dasclaw_runtime           # 64/64 PASS (was 46 在 PR 4a)
cargo fmt --all                                 # OK
cargo clippy --no-deps -p dasclaw_runtime \
    --all-targets -- -D warnings               # OK
python3.12 scripts/check_no_panics.py \
    --base origin/xClaw                        # OK
python3.12 scripts/check_no_new_ironclaw_literal.py \
    --base origin/xClaw                        # OK
```

测试数累计变化:
- PR 3 合并后:19
- PR 4a 合并后:46(+27 types/store)
- 本 PR 后:64(+18 crypto)

## Sources read

- `AGENTS.md` §3(本地必跑管线)、§5(测试纪律 Fail-Safe / req_ 命名)
- `desktop-client/ironclaw/src/secrets/crypto.rs`(被搬的源文件)
- `desktop-client/ironclaw/src/secrets/keychain.rs`(确认有 `SERVICE_NAME="ironclaw"` 硬编码 →
  本 PR 故意不搬)
- `desktop-client/ironclaw/src/secrets/mod.rs`(确认 re-export 路径不变)
- `crates/dasclaw_runtime/src/secrets/mod.rs`(PR 4a 创建,本 PR 修改)
- `docs/plans/architecture-refactor/adr-153-headless-agent-framework-draft.md`
- `docs/plans/architecture-refactor/adr-112-compatibility-evaluation.md`
- `docs/plans/architecture-refactor/adr-113-hook-engine-unification.md`

## 工具协助(self-review 3 层产出)

- **code-quality-audit**:零 `unwrap` / `expect` / `panic!` 在生产代码
  (只在 `#[cfg(test)]` 出现,正常)。`check_no_panics.py` 已校验。
- **code-simplifier**:文件是 verbatim 搬迁,无新增重复代码。
  `format!("{:?}, secret")` → `format!("{secret:?}")` 这种 inline-args
  在搬迁过程中顺手做了(clippy 红线 `uninlined_format_args`)。
- **code-review-expert**:
  - **公开 API 不变**:`crate::secrets::SecretsCrypto` 和
    `crate::secrets::crypto::SecretsCrypto` 两条路径都仍可用
    (前者通过 ironclaw mod.rs 的 `pub use`,后者通过 ironclaw
    `crypto.rs` shim 的 `pub use`)
  - **安全 invariant 保留**:HKDF info 串 `"near-agent-secrets-v1"` 一字不改
    →与已加密数据兼容
  - **新增的安全测试**显式覆盖 Fail-Safe(篡改密文/nonce、错 salt、错
    主密钥),无 Fail-Open 路径
  - **依赖审查**:`aes-gcm 0.10` / `hkdf 0.12` / `sha2 0.10` / `rand 0.8`
    与 ironclaw Cargo.toml 完全一致,避免双版本

## stacked PR 说明

- 本 PR base 不是 `xClaw`,**是 `feat/f3.2-phase2-pr4a-dasclaw-runtime-secrets-trait-#641`**(PR #646)
- 合并顺序:**先合 #646(PR 4a),再合本 PR**
- review 顺序建议:**先看 #646,再看本 PR**
- 本 PR 在 GitHub 上的 diff 默认对比 base 分支,只显示 crypto 相关改动
