+++ crates/dasclaw_runtime/Cargo.toml
@@ -13,6 +13,10 @@ chrono = { version = "0.4", features = ["serde"] }
 secrecy = { version = "0.10", features = ["serde"] }
 uuid = { version = "1", features = ["v4", "serde"] }
 async-trait = "0.1"
+aes-gcm = "0.10"
+hkdf = "0.12"
+sha2 = "0.10"
+rand = "0.8"
 
 [dev-dependencies]
 serde_json = "1"
